+++ src/main/java/fll/web/api/challengeDescription/SubjectiveCategories.java
@@ -19,6 +19,7 @@ import javax.servlet.http.HttpServletResponse;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import fll.Utilities;
 import fll.web.ApplicationAttributes;
 import fll.xml.ChallengeDescription;
 import fll.xml.SubjectiveScoreCategory;
@@ -35,7 +36,7 @@ public class SubjectiveCategories extends HttpServlet {
     response.setContentType("application/json");
     final PrintWriter writer = response.getWriter();
 
-    final ObjectMapper jsonMapper = new ObjectMapper();
+    final ObjectMapper jsonMapper = Utilities.createJsonMapper();
 
     final ChallengeDescription challengeDescription = ApplicationAttributes.getChallengeDescription(application);
     final List<SubjectiveScoreCategory> categories = challengeDescription.getSubjectiveCategories();